+++ cucumber_tests/framework/golden_compare.py
