--- dependencies/pip.conf
@@ -1,1 +1,0 @@
-
