--- dependencies/sources.list
@@ -1,1 +1,0 @@
-
