+++ devtools/cucumber/cucumber_requirements.txt
@@ -1,4 +1,3 @@
 pytest==8.3.5
 pytest-bdd
 junitparser
-cryptography==46.0.5
