+++ devtools/local_runner/Dockerfile
@@ -12,13 +12,8 @@ RUN curl -LO https://github.com/getsops/sops/releases/download/v3.9.0/sops-v3.9.
     mv sops-v3.9.0.linux.amd64 /usr/local/bin/sops && \
     chmod +x /usr/local/bin/sops
 
-# Optional pip.conf / sources.list
-COPY dependencies/pip.conf /etc/pip.conf
-COPY dependencies/sources.list /etc/apt/sources.list
-
 # PyPI test dependencies only (repo packages are installed editable at container start via up.sh)
-COPY dependencies/tests_requirements.txt /tmp/tests_requirements.txt
+COPY devtools/cucumber/cucumber_requirements.txt /tmp/cucumber_requirements.txt
 # hadolint ignore=DL3013,DL3042
 RUN pip install --no-cache-dir "uv>=0.9.5" && \
-    uv pip install --system --upgrade pip "setuptools<82" wheel && \
-    uv pip install --system --no-cache-dir -r /tmp/tests_requirements.txt
+    uv pip install --system --no-cache-dir -r /tmp/cucumber_requirements.txt
