+++ devtools/local_runner/run.sh
@@ -20,11 +20,6 @@ run_pytest_suite() {
 }
 
 run_pytest_suite envgenehelper python/envgene/envgenehelper
-run_pytest_suite pipegene build_pipegene/scripts
-run_pytest_suite artifact_searcher python/artifact-searcher/artifact_searcher
-run_pytest_suite bg_manage scripts/bg_manage
-run_pytest_suite build_env scripts/build_env
-run_pytest_suite cred_rotation creds_rotation/scripts
-run_pytest_suite sbom_retention build_effective_set_generator/scripts
+run_pytest_suite scripts scripts/tests
 
 junitparser merge junit_*.xml junit.xml
