--- pytest.ini
@@ -1,7 +1,0 @@
-[pytest]
-testpaths = tests/step_defs
-python_files = test_*.py
-python_classes = Test*
-python_functions = test_*
-addopts = -v
-pythonpath = .
