--- tests/step_defs/common_steps.py
@@ -1,1 +1,0 @@
-from tests.shared_steps.common_steps import *
