--- tests/step_defs/test_environment_inventory_generation.py
@@ -1,5 +1,0 @@
-from pytest_bdd import scenarios
-from tests.shared_steps.inventory_gen_steps import *
-from tests.shared_steps.common_steps import *
-
-scenarios('../features/environment-inventory-generation.feature')
